Minor fixes, comment, and code cleanup.
So much cleanup. Strike Team handling is very solid now. Always generates a strike_teams.py file if none present in the current directory and uses this info for subsequent runs. This works out great for cases where the end user is running a frozen EXE on their machine -- they can edit this file with new names or orders and we will use that file for output.