Per alliance settings, grayscale = stale data
Added support for strike_teams.py and raids_and_lanes.py on an alliance by alliance basis. 
To use, create a folder with the same name as the Alliance and put a copy of strike_teams.py and raids_and_lanes.py inside it to use with this alliance.
Also note that .msf file inside the folder will be found and output will also be placed in this directory.
Finally, changed tables so that if roster hasn't been re-synced in more than a week, output is changed to grayscale to indicate stale data is in use.